Move rules service to shared
