+++ app/api/api/views.py
@@ -61,12 +61,6 @@ from mapping.permissions import (
     get_user_permissions_on_scan_report,
 )
 from mapping.services.files import delete_blob, modify_filename, upload_blob
-from mapping.services.rules import (
-    get_mapping_rules_as_csv,
-    get_mapping_rules_json,
-    get_mapping_rules_list,
-    make_dag,
-)
 from rest_framework import generics, status, viewsets
 from rest_framework.filters import OrderingFilter
 from rest_framework.generics import ListAPIView, RetrieveAPIView
@@ -96,6 +90,12 @@ from shared.services.rules import (
     _find_destination_table,
     _save_mapping_rules,
     delete_mapping_rules,
+)
+from shared.services.rules_export import (
+    get_mapping_rules_as_csv,
+    get_mapping_rules_json,
+    get_mapping_rules_list,
+    make_dag,
 )
