+++ app/api/mapping/views.py
@@ -36,10 +36,14 @@ from shared.data.models import (
     ScanReportTable,
 )
 from shared.services.azurequeue import add_message
+from shared.services.rules_export import (
+    get_mapping_rules_as_csv,
+    get_mapping_rules_json,
+    make_dag,
+)
 
 from .forms import ScanReportAssertionForm, ScanReportForm
 from .permissions import has_editorship, has_viewership, is_admin
-from .services.rules import get_mapping_rules_as_csv, get_mapping_rules_json, make_dag
 
 
 @login_required
